devops-cpu-fix-demo/service/requirements.txt
fastapi
uvicorn
python-dotenv
prometheus-client
requests
